+++ app/src/main/java/woowacourse/shopping/presentation/view/order/cart/adapter/CartViewHolder.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.presentation.view.cart.adapter
+package woowacourse.shopping.presentation.view.order.cart.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
