+++ app/src/main/java/woowacourse/shopping/presentation/view/order/suggestion/adapter/SuggestionProductViewHolder.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.presentation.view.cart.adapter
+package woowacourse.shopping.presentation.view.order.suggestion.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
