+++ app/src/main/java/woowacourse/shopping/presentation/view/order/suggestion/event/SuggestionMessageEvent.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.presentation.view.cart.event
+package woowacourse.shopping.presentation.view.order.suggestion.event
 
 enum class SuggestionMessageEvent {
     FETCH_SUGGESTION_PRODUCT_FAILURE,
