+++ app/src/main/res/layout/fragment_cart.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="vm"
-            type="woowacourse.shopping.presentation.view.cart.OrderViewModel" />
+            type="woowacourse.shopping.presentation.view.order.OrderViewModel" />
 
     </data>
