+++ app/src/main/res/layout/item_cart.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="eventListener"
-            type="woowacourse.shopping.presentation.view.cart.adapter.CartAdapter.CartEventListener" />
+            type="woowacourse.shopping.presentation.view.order.cart.adapter.CartAdapter.CartEventListener" />
 
     </data>
