+++ app/src/main/res/layout/item_suggestion.xml
@@ -17,7 +17,7 @@
 
         <variable
             name="eventListener"
-            type="woowacourse.shopping.presentation.view.cart.adapter.SuggestionAdapter.SuggestionEventListener" />
+            type="woowacourse.shopping.presentation.view.order.suggestion.adapter.SuggestionAdapter.SuggestionEventListener" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
